// Appraisal-Teaching-Processes/dict/api/teaching/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Appraisal Service
 * Appraisal Service
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface TeachingApi {
	postTeachingProcessCreate: (request: Api.TeachingData | undefined) => Promise<PostTeachingProcessCreateResponse>
	deleteTeachingProcessDelete: (id: string) => Promise<DeleteTeachingProcessDeleteResponse>
	getTeachingProcessGet: (id: string) => Promise<GetTeachingProcessGetResponse>
	getTeachingProcessGetAll: () => Promise<GetTeachingProcessGetAllResponse>
	putTeachingProcessUpdate: (id: string, request: Api.TeachingData | undefined) => Promise<PutTeachingProcessUpdateResponse>
}

export type PostTeachingProcessCreateResponse = PostTeachingProcessCreate201Response | PostTeachingProcessCreate401Response | PostTeachingProcessCreate404Response | PostTeachingProcessCreate422Response

export interface PostTeachingProcessCreate201Response {
	status: 201
	body: Api.TeachingData
	headers?: never
}

export interface PostTeachingProcessCreate401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface PostTeachingProcessCreate404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface PostTeachingProcessCreate422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}

export type DeleteTeachingProcessDeleteResponse = DeleteTeachingProcessDelete200Response | DeleteTeachingProcessDelete401Response | DeleteTeachingProcessDelete404Response | DeleteTeachingProcessDelete422Response

export interface DeleteTeachingProcessDelete200Response {
	status: 200
	body: Api.MessageData
	headers?: never
}

export interface DeleteTeachingProcessDelete401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface DeleteTeachingProcessDelete404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface DeleteTeachingProcessDelete422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}

export type GetTeachingProcessGetResponse = GetTeachingProcessGet200Response | GetTeachingProcessGet401Response | GetTeachingProcessGet404Response | GetTeachingProcessGet422Response

export interface GetTeachingProcessGet200Response {
	status: 200
	body: Api.TeachingData
	headers?: never
}

export interface GetTeachingProcessGet401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface GetTeachingProcessGet404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface GetTeachingProcessGet422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}

export type GetTeachingProcessGetAllResponse = GetTeachingProcessGetAll200Response | GetTeachingProcessGetAll401Response | GetTeachingProcessGetAll404Response

export interface GetTeachingProcessGetAll200Response {
	status: 200
	body: Api.TeachingPagedResultData
	headers?: never
}

export interface GetTeachingProcessGetAll401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface GetTeachingProcessGetAll404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export type PutTeachingProcessUpdateResponse = PutTeachingProcessUpdate200Response | PutTeachingProcessUpdate401Response | PutTeachingProcessUpdate404Response | PutTeachingProcessUpdate422Response

export interface PutTeachingProcessUpdate200Response {
	status: 200
	body: Api.TeachingData
	headers?: never
}

export interface PutTeachingProcessUpdate401Response {
	status: 401
	body: Api.MessageData
	headers?: never
}

export interface PutTeachingProcessUpdate404Response {
	status: 404
	body: Api.MessageData
	headers?: never
}

export interface PutTeachingProcessUpdate422Response {
	status: 422
	body: Api.MessageData
	headers?: never
}